Change CString::Find return type from int to size_t and fix build warnings
Return size_t from CString::Find (both overloads) to match std::string::find
semantics, eliminating signed/unsigned comparison issues across the codebase.
All ~30 call sites updated to use string::npos instead of -1 or >= 0 checks.

Also fix 6 deprecation/tautology warnings in library code:
- sprintf/vsprintf → snprintf/vsnprintf (Thread.cpp, CException.cpp, Digest.cpp)
- Remove tautological _data != NULL check (DataBuffer.cpp)
- Remove unused _network_id field (GenericServer.h)

All 552 tests pass (JTC: 152, EIBStdLib: 272, EIBServer unit: 93,
EIBServer integration: 35).

+++ EIBStdLib/src/CString.cpp
@@ -293,9 +293,9 @@ int CString::FindFirstOf(const CString& c) const
 	return ret;
 }
 
-int CString::Find(char c, int pos) const
+size_t CString::Find(char c, size_t pos) const
 {
-	return (int)_str.find(c,pos);
+	return _str.find(c,pos);
 }
 
 CString CString::SubString(int start,int length)  const
@@ -305,9 +305,9 @@ CString CString::SubString(int start,int length)  const
 	return ret;
 }
 
-int CString::Find(const CString& value, int pos) const
+size_t CString::Find(const CString& value, size_t pos) const
 {
-	return (int)_str.find(value._str.c_str(),pos);
+	return _str.find(value._str.c_str(),pos);
 }
 
 int CString::RFind(const CString& value, int pos) const
